requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.1.5
django-bootstrap-form==3.4
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.38.0
gunicorn==20.1.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
openpyxl==3.0.10
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.3
static3==0.7.0
